+++ examples/foxes.opt/layout_pymoo/run.py
@@ -1,5 +1,69 @@
+import numpy as np
+import argparse
+
 import foxes
 from foxes.opt import FarmProblem
 
 if __name__ == "__main__":
-    pass
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-nt", "--n_t", help="The number of turbines", type=int, default=5
+    )
+    parser.add_argument(
+        "-t",
+        "--turbine_file",
+        help="The P-ct-curve csv file (path or static)",
+        default="NREL-5MW-D126-H90.csv",
+    )
+    parser.add_argument("-r", "--rotor", help="The rotor model", default="centre")
+    parser.add_argument(
+        "-w",
+        "--wakes",
+        help="The wake models",
+        default=["Jensen_linear_k007"],
+        nargs="+",
+    )
+    parser.add_argument(
+        "-p", "--pwakes", help="The partial wakes model", default="rotor_points"
+    )
+    parser.add_argument("--ws", help="The wind speed", type=float, default=9.0)
+    parser.add_argument("--wd", help="The wind direction", type=float, default=270.0)
+    parser.add_argument("--ti", help="The TI value", type=float, default=0.08)
+    parser.add_argument("--rho", help="The air density", type=float, default=1.225)
+    parser.add_argument(
+        "-c", "--chunksize", help="The maximal chunk size", type=int, default=1000
+    )
+    args = parser.parse_args()
+
+    mbook = foxes.models.ModelBook()
+    ttype = foxes.models.turbine_types.PCtFile(args.turbine_file)
+    mbook.turbine_types[ttype.name] = ttype
+
+    farm = foxes.WindFarm()
+    foxes.input.farm_layout.add_row(
+        farm=farm,
+        xy_base=np.zeros(2),
+        xy_step=np.array([500., 0.]),
+        n_turbines=args.n_t,
+        turbine_models=[ttype.name],
+    )
+
+    states = foxes.input.states.SingleStateStates(
+        ws=args.ws, wd=args.wd, ti=args.ti, rho=args.rho
+    )
+
+    algo = foxes.algorithms.Downwind(
+        mbook,
+        farm,
+        states=states,
+        rotor_model=args.rotor,
+        wake_models=args.wakes,
+        wake_frame="rotor_wd",
+        partial_wakes_model=args.pwakes,
+        chunks=args.chunksize,
+    )
+    algo.initialize()
+
+    runner = foxes.models.runners.DaskRunner(n_workers=2, verbosity=1)
+