+++ foxes/models/__init__.py
@@ -9,5 +9,6 @@ from . import wake_superpositions
 from . import farm_controllers
 from . import vertical_profiles
 from . import point_models
+from . import runners
 
 from .model_book import ModelBook