+++ foxes/models/runners/__init__.py
@@ -1,0 +1,2 @@
+from .simple_runner import SerialRunner
+from .dask_runner import DaskRunner